Add CSS variable for font fallback of specimen font(s)
This way a specimen author can set a variable to use in the
specimen font stack. This can be used to catch characters that
somehow aren't in the specimen's font. (E.g. when a user types
a character that's not supported)

An author could decide to show the unsupported characters in a font
like Adobe NotDef (https://github.com/adobe-fonts/adobe-notdef/)

Note that monospace is added twice: once as fallback for when
the CSS variable is undefined, once as ultimate fallback (mostly
to satisfy the linter).